+++ src/services/nostr.js
@@ -16,9 +16,7 @@ let publicKey;
 let encryptionPrivateKey; // Added for content encryption
 let encryptionPublicKey; // Added for content encryption
 
-// Configuration for developer fee
-const DEV_FEE_PERCENT = 5; // Default 5% developer fee
-const DEV_PUBKEY = 'a745806d90a71d89f1a33ed9c349834c45ae4c071493639afd3d25e4f411a0a5'; // Developer public key
+const DEV_PUBKEY = 'a745806d90a71d89f1a33ed9c349834c45ae4c071493639afd3d25e4f411a0a5'; // Developer public key (Needed for DM target)
 
 // Connect to relays
 const connect = async () => {
@@ -61,116 +59,6 @@ const initializeKeys = ()  => {
 };
 
 /**
- * Creates a standard Cashu payment request without structured token denominations
- * @param {Number} totalAmount - Total payment amount in sats
- * @returns {String} NUT-18 Cashu payment request
- */
-const createPaymentRequest = async (totalAmount) => {
-  // This would call your actual Cashu library's function to create a standard payment request
-  // Implementation depends on your Cashu library
-  return generateCashuPaymentRequest(totalAmount);
-};
-
-/**
- * Processes received tokens and splits them between author and developer
- * @param {Object} receivedToken - The token received from payment
- * @param {String} authorWallet - Author's wallet address
- * @param {Number} devFeePercent - Developer fee percentage
- * @returns {Object} Result of payment splitting
- */
-const processSplitPayment = async (receivedToken, authorWallet, devFeePercent = DEV_FEE_PERCENT) => {
-  try {
-    // Get total amount from the received token's proofs
-    const totalAmount = receivedToken.proofs.reduce((sum, proof) => sum + proof.amount, 0);
-    
-    // Calculate the split (same total, just divided according to percentages)
-    const devFeeAmount = Math.floor(totalAmount * (devFeePercent / 100));
-    const authorAmount = totalAmount - devFeeAmount;
-    
-    console.log(`Splitting payment: ${authorAmount} to receipt creator (${100-devFeePercent}%), ${devFeeAmount} to developer (${devFeePercent}%)`);
-    
-    // Split the proofs based on amounts
-    const { authorProofs, devProofs } = splitProofs(receivedToken.proofs, devFeeAmount);
-    
-    // Create tokens with the split proofs
-    const authorToken = { ...receivedToken, proofs: authorProofs };
-    const devToken = { ...receivedToken, proofs: devProofs };
-    // Send tokens to respective wallets
-    const authorResult = await sendTokenToWallet(authorToken, authorWallet);
-    const devResult = await sendTokenToWallet(devToken, DEV_PUBKEY);
-    
-    // Send NIP-04 DM to developer with payment details
-    if (devProofs.length > 0) {
-      await sendDeveloperPaymentNotification(devFeeAmount, authorWallet);
-    }
-    
-    return {
-      success: true,
-      authorAmount,
-      devFeeAmount,
-      authorResult,
-      devResult
-    };
-  } catch (error) {
-    console.error('Error processing split payment:', error);
-    throw new Error(`Failed to split payment: ${error.message}`);
-  }
-};
-
-/**
- * Splits proofs within a Cashu token to separate developer fee from author amount
- * @param {Array} proofs - Array of proof objects from the token
- * @param {Number} devFeeAmount - Amount to allocate to developer
- * @returns {Object} Object containing authorProofs and devProofs arrays
- */
-const splitProofs = (proofs, devFeeAmount) => {
-  // Sort proofs by amount (small to large)
-  const sortedProofs = [...proofs].sort((a, b) => a.amount - b.amount);
-  
-  const devProofs = [];
-  const authorProofs = [];
-  let devTotal = 0;
-  
-  // First pass: try to find exact proofs for dev fee
-  for (let i = 0; i < sortedProofs.length; i++) {
-    if (sortedProofs[i].amount === devFeeAmount - devTotal) {
-      devProofs.push(sortedProofs[i]);
-      devTotal += sortedProofs[i].amount;
-      sortedProofs.splice(i, 1);
-      break;
-    }
-  }
-  
-  // Second pass: accumulate proofs until we reach dev fee amount
-  if (devTotal < devFeeAmount) {
-    for (let i = 0; i < sortedProofs.length; i++) {
-      if (devTotal + sortedProofs[i].amount <= devFeeAmount) {
-        devProofs.push(sortedProofs[i]);
-        devTotal += sortedProofs[i].amount;
-        sortedProofs.splice(i, 1);
-        i--;
-      }
-    }
-  }
-  
-  // If we still need more for dev fee, we'll need to split a proof
-  if (devTotal < devFeeAmount && sortedProofs.length > 0) {
-    // This would call your Cashu library's function to split a proof
-    const remaining = devFeeAmount - devTotal;
-    const { smallerProof, largerProof } = splitProof(sortedProofs[0], remaining);
-    
-    devProofs.push(smallerProof);
-    authorProofs.push(largerProof);
-    sortedProofs.splice(0, 1);
-  }
-  
-  // Remaining proofs go to author
-  authorProofs.push(...sortedProofs);
-  
-  return { authorProofs, devProofs };
-};
-
-/**
  * Publish a receipt event (kind 9567)
  * @param {Object} receiptData - The receipt data in JSON format
  * @param {String} paymentRequest - The NUT-18 Cashu payment request or LNURL
@@ -178,7 +66,7 @@ const splitProofs = (proofs, devFeeAmount) => {
  * @param {Number} devFeePercent - Developer fee percentage
  * @returns {Object} The event ID and encryption key
  */
-const publishReceiptEvent = async (receiptData, paymentRequest, paymentType, devFeePercent = DEV_FEE_PERCENT) => {
+const publishReceiptEvent = async (receiptData, paymentRequest, paymentType, devFeePercent) => {
   try {
     // Initialize keys and connect if not already done
     const { privateKey: pk, publicKey: pubKey, encryptionPrivateKey: encryptionPrivateKey, encryptionPublicKey: encryptionPublicKey } = initializeKeys();
@@ -235,10 +123,12 @@ const publishReceiptEvent = async (receiptData, paymentRequest, paymentType, dev
  * @param {Number} devFeePercent - Developer fee percentage
  * @returns {Object} The receipt and payment details
  */
-const processPaymentWithSplit = async (receiptData, authorWallet, devFeePercent = DEV_FEE_PERCENT) => {
-  try {
-    // Create a standard payment request (not structured)
-    const paymentRequest = await createPaymentRequest(receiptData.amount);
+const processPaymentWithSplit_UNUSED = async (receiptData, authorWallet, devFeePercent = 5 /* Default 5% */) => {
+  // NOTE: This function seems unused and relies on the now-moved createPaymentRequest.
+  // Keeping it commented out or removing it might be best.
+  // try {
+  //   // Create a standard payment request (not structured)
+  //   // const paymentRequest = await cashuService.createPaymentRequest(receiptData.amount); // Needs cashuService import
     
     // Calculate the expected split (for information only)
     const devFeeAmount = Math.floor(receiptData.amount * (devFeePercent / 100));
@@ -264,53 +154,6 @@ const processPaymentWithSplit = async (receiptData, authorWallet, devFeePercent 
       // Add a note for processing required
       processingRequired: true
     };
-  } catch (error) {
-    console.error('Error processing payment with split:', error);
-    throw new Error(`Failed to process payment: ${error.message}`);
-  }
-};
-
-/**
- * Handle incoming payment and process the split
- * @param {String} paymentId - ID of the payment
- * @param {Object} receivedToken - Token received from payment
- * @param {String} authorWallet - Author's wallet address
- * @param {Number} devFeePercent - Developer fee percentage
- * @param {String} payerRequest - Original payer's payment request to forward funds
- * @returns {Object} Result of processing
- */
-const handlePaymentReceived = async (paymentId, receivedToken, authorWallet, devFeePercent = DEV_FEE_PERCENT, payerRequest = null) => {
-  try {
-    // Process the split payment
-    const splitResult = await processSplitPayment(receivedToken, authorWallet, devFeePercent);
-    
-    // If there's a payer request, forward the author's share to it
-    let forwardResult = null;
-    if (payerRequest && splitResult.authorProofs && splitResult.authorProofs.length > 0) {
-      // Create a token with the author's proofs
-      const authorToken = { ...receivedToken, proofs: splitResult.authorProofs };
-      
-      // Forward payment to the original payer's request
-      forwardResult = await forwardPaymentToRequest(authorToken, payerRequest);
-      
-      console.log(`Forwarded ${splitResult.authorAmount} sats to original payer's request`);
-    }
-    
-    // Update payment status in your database or storage
-    // This would depend on your application's storage implementation
-    // updatePaymentStatus(paymentId, 'completed', splitResult);
-    
-    return {
-      success: true,
-      paymentId,
-      ...splitResult,
-      forwarded: forwardResult ? true : false,
-      forwardResult
-    };
-  } catch (error) {
-    console.error('Error handling received payment:', error);
-    throw new Error(`Failed to process received payment: ${error.message}`);
-  }
 };
 
 /**
@@ -492,7 +335,7 @@ const subscribeToPaymentUpdates = async (paymentId, callback) => {
  * @param {String} fromWallet - Wallet address of the payer
  * @returns {Promise} Promise that resolves when message is sent
  */
-const sendDeveloperPaymentNotification = async (amount, fromWallet) => {
+const sendDeveloperPaymentNotification = async (amount, fromWallet, percentage) => {
   try {
     if (!ndk.pool?.connectedRelays?.size) {
       await connect();
@@ -503,7 +346,7 @@ const sendDeveloperPaymentNotification = async (amount, fromWallet) => {
       type: 'developer_fee',
       amount,
       fromWallet,
-      percentage: DEV_FEE_PERCENT,
+      percentage: percentage, // Use passed percentage
       timestamp: Math.floor(Date.now() / 1000)
     };
     
@@ -526,41 +369,6 @@ const sendDeveloperPaymentNotification = async (amount, fromWallet) => {
   }
 };
 
-/**
- * Forward payment to the original payer's payment request
- * @param {Object} token - The token to forward
- * @param {String} paymentRequest - The payment request to send funds to
- * @returns {Object} Result of the forwarding operation
- */
-const forwardPaymentToRequest = async (token, paymentRequest) => {
-  try {
-    // In a real implementation, this would use the Cashu library to
-    // send the token to the payment request
-    
-    // For demonstration purposes:
-    console.log(`Forwarding payment to request: ${paymentRequest}`);
-    console.log(`Token proofs to forward: ${token.proofs.length}`);
-    
-    // This would call your Cashu library's function to redeem a token to a payment request
-    // const result = await cashuLibrary.redeemToken(token, paymentRequest);
-    
-    // Simulate a successful result
-    const result = {
-      success: true,
-      request: paymentRequest,
-      amount: token.proofs.reduce((sum, proof) => sum + proof.amount, 0)
-    };
-    
-    return result;
-  } catch (error) {
-    console.error('Error forwarding payment:', error);
-    return {
-      success: false,
-      error: error.message
-    };
-  }
-};
-
 // Export the service functions
 export default {
   connect,
@@ -569,11 +377,11 @@ export default {
   publishSettlementEvent,
   subscribeToSettlements,
   fetchReceiptEvent,
-  createPaymentRequest,
-  processSplitPayment,
-  processPaymentWithSplit,
-  handlePaymentReceived,
-  sendDeveloperPaymentNotification,
+  // createPaymentRequest, // Moved
+  // processSplitPayment, // Moved
+  // processPaymentWithSplit, // Unused / Relies on moved functions
+  // handlePaymentReceived, // Moved
+  sendDeveloperPaymentNotification, // Kept
   subscribeToPaymentUpdates,
-  forwardPaymentToRequest
+  // forwardPaymentToRequest // Moved
 };
